Refactor groupedInputFields
