+++ website/src/pages/[organism]/submission/[groupId]/revise.astro
@@ -2,9 +2,8 @@
 import { cleanOrganism } from '../../../../components/Navigation/cleanOrganism';
 import { RevisionForm } from '../../../../components/Submission/RevisionForm';
 import SubmissionPageWrapper from '../../../../components/Submission/SubmissionPageWrapper.astro';
-import { getRuntimeConfig, getSchema } from '../../../../config';
+import { getGroupedInputFields, getRuntimeConfig } from '../../../../config';
 import { getAccessToken } from '../../../../utils/getAccessToken';
-import { groupFieldsByHeader } from '../../../../utils/groupFieldsByHeader';
 import { getReferenceGenomesSequenceNames } from '../../../../utils/search';
 import { getGroupsAndCurrentGroup } from '../../../../utils/submissionPages';
 
@@ -18,9 +17,7 @@ if (!cleanedOrganism) {
     };
 }
 const referenceGenomeSequenceNames = getReferenceGenomesSequenceNames(cleanedOrganism.key);
-
-const schema = getSchema(cleanedOrganism.key);
-const groupedInputFields = groupFieldsByHeader(schema.inputFields, schema.metadata, 'revise');
+const groupedInputFields = getGroupedInputFields(cleanedOrganism.key, 'revise');
 
 const groupsResult = await getGroupsAndCurrentGroup(Astro.params, Astro.locals.session);
